chore: rename & move Product Controllers
Rename: ProductRestController to ProductController
Move: ServerRenderController to gift.serverside.rendering.ProductController
